requirements.txt
pytz>=2019.3
torch>=1.3.0
tensorboard>=2.1.0
tensorboardX>=2.0
sympy>=1.5.1
pybind11>=2.4.dev4
cython>=0.29.14
numpy>=1.18.1
numpy-quaternion>=2019.12.11.22.25.52
numba>=0.47.0
scipy>=1.4.1
scikit-image>=0.16.2
colorama>=0.4.3
setproctitle>=1.1.10
tqdm>=4.41.1
cffi>=1.13.2
Pillow>=7.0.0
